Add remove function
